pyproject.toml
[project]
name = "github-to-markdown-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
    "uvicorn[standard]>=0.34.2",
]
